public-api/requirements.txt
fastapi
uvicorn
gunicorn
databases[postgresql]
python-dotenv
psycopg2
fastapi-cache2[redis]
redis
